fix(ci): python-full-coverage — split oversized part_007 and close remaining residue fail-closed gaps
